flask_server/requirements.txt
click==8.1.3
confluent-kafka==1.9.0
Flask==2.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.20.0
Werkzeug==2.1.2
